Implement authorization checks for post creation, update, and deletion

+++ app/Http/Controllers/Api/PostController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
-use App\Http\Requests\StorePostRequest;
-use App\Http\Resources\PostResource;
 use App\Models\Post;
+use App\Http\Controllers\Controller;
+use App\Http\Resources\PostResource;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\StorePostRequest;
 use Illuminate\Contracts\Database\Eloquent\Builder;
 
 class PostController extends Controller
@@ -50,6 +51,8 @@ class PostController extends Controller
 
     public function store(StorePostRequest $request)
     {
+        Gate::authorize('posts.create'); 
+
         if ($request->hasFile('thumbnail')) {
             $filename = $request->file('thumbnail')->getClientOriginalName();
             info($filename);
@@ -62,11 +65,15 @@ class PostController extends Controller
 
     public function show(Post $post)
     {
+        Gate::authorize('posts.update'); 
+
         return new PostResource($post);
     }
 
     public function update(Post $post, StorePostRequest $request)
     {
+        Gate::authorize('posts.update'); 
+
         $post->update($request->validated());
 
         return new PostResource($post);
@@ -74,6 +81,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        Gate::authorize('posts.delete'); 
+        
         $post->delete();
 
         return response()->noContent();
